# triton-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Triton Backend",
    "description": "Backend for Triton",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/users/{user_id}/instances": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update User Instances",
        "operationId": "update_user_instances_api_auth_users__user_id__instances_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInstancesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instances": {
      "post": {
        "tags": [
          "instances"
        ],
        "summary": "Create Instance",
        "description": "Create a Triton instance from UI input (url + optional name).\n\nThe backend connects to Triton, reads the currently available models,\nstores the instance in the DB, and returns the created DTO.",
        "operationId": "create_instance_api_instances_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTritonInstanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TritonInstanceDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "List Instances",
        "description": "List all registered Triton instances.",
        "operationId": "list_instances_api_instances_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TritonInstanceDTO"
                  },
                  "title": "Response List Instances Api Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance",
        "description": "Get a single Triton instance by ID.",
        "operationId": "get_instance_api_instances__instance_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TritonInstanceDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "instances"
        ],
        "summary": "Delete Instance",
        "description": "Delete a registered Triton instance. Admin role required.",
        "operationId": "delete_instance_api_instances__instance_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "instances"
        ],
        "summary": "Update Instance",
        "description": "Update a Triton instance endpoint and SSL verification settings.",
        "operationId": "update_instance_api_instances__instance_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTritonInstanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TritonInstanceDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/by-name/{name}": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance By Name",
        "description": "Get a single Triton instance by its unique name.",
        "operationId": "get_instance_by_name_api_instances_by_name__name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TritonInstanceDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTritonInstanceRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "verify_ssl": {
            "type": "boolean",
            "title": "Verify Ssl",
            "default": false
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CreateTritonInstanceRequest",
        "description": "Request to create a Triton instance."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstanceS3ConfigDTO": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefix"
          },
          "access_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key"
          },
          "secret_configured": {
            "type": "boolean",
            "title": "Secret Configured",
            "default": false
          },
          "use_https": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Https"
          },
          "verify_ssl": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Ssl"
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "address_style": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "path",
                  "virtual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Style"
          }
        },
        "type": "object",
        "title": "InstanceS3ConfigDTO"
      },
...
      "TritonInstanceDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "model_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Model Names"
          },
          "repository_models": {
            "items": {
              "$ref": "#/components/schemas/TritonRepositoryModelDTO"
            },
            "type": "array",
            "title": "Repository Models"
          },
          "server_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Metadata"
          },
          "health_live": {
            "type": "boolean",
            "title": "Health Live",
            "default": false
          },
          "health_ready": {
            "type": "boolean",
            "title": "Health Ready",
            "default": false
          },
          "health_last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Last Checked At"
          },
          "health_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Error"
          },
          "triton_verify_ssl": {
            "type": "boolean",
            "title": "Triton Verify Ssl",
            "default": false
          },
          "triton_ca_certificate": {
            "type": "string",
            "title": "Triton Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          },
          "metrics_cpu": {
            "type": "number",
            "title": "Metrics Cpu",
            "default": 0
          },
          "metrics_ram": {
            "type": "number",
            "title": "Metrics Ram",
            "default": 0
          },
          "metrics_gpu": {
            "type": "number",
            "title": "Metrics Gpu",
            "default": 0
          },
          "metrics_last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Last Checked At"
          },
          "metrics_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Error"
          },
          "deployment_runtime": {
            "type": "string",
            "title": "Deployment Runtime",
            "default": "external"
          },
          "deployment_namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Namespace"
          },
          "deployment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Name"
          },
          "deployment_service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Service Name"
          },
          "deployment_secret_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Secret Name"
          },
          "deployment_log": {
            "type": "string",
            "title": "Deployment Log",
            "default": ""
          },
          "is_self_deployed": {
            "type": "boolean",
            "title": "Is Self Deployed",
            "default": false
          },
          "pod_statuses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pod Statuses"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "s3": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstanceS3ConfigDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "name",
          "model_names",
          "created_at"
        ],
        "title": "TritonInstanceDTO",
        "description": "DTO for outgoing Triton instance data (API responses)."
      },
      "TritonRepositoryModelDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TritonRepositoryModelDTO"
      },
...
      "UpdateTritonInstanceRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "verify_ssl": {
            "type": "boolean",
            "title": "Verify Ssl",
            "default": false
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "UpdateTritonInstanceRequest",
        "description": "Request to update an existing Triton instance connection."
      },
      "UpdateUserInstancesRequest": {
        "properties": {
          "assigned_instances": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assigned Instances"
          }
        },
        "type": "object",
        "title": "UpdateUserInstancesRequest"
      },
...
      "UserDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "auth_provider": {
            "type": "string",
            "title": "Auth Provider"
          },
          "oidc_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oidc Subject"
          },
          "assigned_instances": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assigned Instances"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "auth_provider",
          "assigned_instances",
          "is_active",
          "created_at"
        ],
        "title": "UserDTO"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}